removing shop-product component - changing shop products component
