--- src/app/shop-online/shop-products/shop-product/shop-product.component.html
@@ -1,29 +1,0 @@
-<div *ngFor="let product of shopProducts | paginate : {
-    itemsPerPage: selectedAmount, 
-    currentPage: page, 
-    totalItems: count}; let i = index" class="product" [ngClass]="{list: selectedLayout === 'list'}">
-    <div class="product__content">
-        <h2 class="product__content-title">{{product.name | shorten: 15}}</h2>
-        <p class="product__content-description">{{product.description | shorten: 50}}</p>
-    </div>
-    <div class="product__img">
-        <img class="img" src="{{product.imagePath}}" alt="">
-    </div>
-    <div class="product__summary">
-        <p class="product__summary-price">{{ product.price | currency }}</p>
-        <div class="product__summary-add">
-            <div class="product__summary-add-cart">
-                <p class="product__summary-quantity">{{product.quantity ? product.quantity : ' '}}</p>
-                <i class="product__summary-icon fa-solid fa-cart-shopping" (click)="onAddToCart(product)"></i>
-            </div>
-            <i class="product__summary-icon fa-solid fa-heart" (click)="onAddToFavourites(product)"
-                [ngClass]="{active : onCheckFavourites(product)}"></i>
-        </div>
-    </div>
-</div>
-<app-alert *ngIf="activeAlert"></app-alert>
-<div class="product__pagination">
-    <pagination-controls *ngIf="selectedAmount < shopProducts.length" class="pagination-controls" previouslabel='Prev'
-        nextLabel='next' (pageChange)="onTableDataChange($event)">
-    </pagination-controls>
-</div>
