--- src/app/shop-online/shop-products/shop-product/shop-product.component.scss
@@ -1,193 +1,0 @@
-.product {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    margin: 0.5em;
-    padding: 1em;
-    width: 100%;
-    border: thin solid rgb(206, 206, 206);
-    border-radius: 5px;
-
-    &__content {
-        height: 30%;
-        
-        &-title {
-            padding-bottom: 0.7em;
-            font-size: 1.6rem;
-            font-weight: bold;
-            color: grey;
-        }
-
-        &-description {
-            font-size: 1.2rem;
-        }
-    }
-
-    &__img {
-        width: 100%;
-        height: 50%;
-        
-        & .img {
-            display: block;
-            margin: 0 auto;
-            padding: 1em;
-            width: 80%;
-            height: 100%;
-            filter: grayscale(0.5);
-        }
-    }
-
-    &__summary {
-        display: flex;
-        justify-content: space-between;
-        align-items: flex-end;
-        padding-top: 2em;
-        height: 20%;
-
-        &-add {
-            display: flex;
-            align-items: center;
-
-            &-cart {
-                position: relative;
-            }
-        }
-
-        &-price {
-            margin: 0;
-        }
-
-        &-quantity {
-           position: absolute;
-           bottom: 0;
-           left: 50%;
-           transform: translate(-50%, -50%);
-           display: inline;
-           font-size: 1rem;
-        }
-
-        &-icon {
-            margin: 0 0.2em;
-            font-size: 1.7rem;
-            color: rgb(169, 169, 169);
-            cursor: pointer;
-
-            &:hover,
-            &.active {
-                color: rgb(224, 200, 204);
-            }
-        }
-    }
-
-    &.list {
-        flex-direction: row;
-        flex-wrap: wrap;
-        height: 100%;
-
-        & .product__img {
-            width: 50%;
-            height: 130px;
-
-            & .img {
-                padding: 0 0 0 1em;
-                width: 100%;
-                height: 100%;
-                filter: grayscale(0.5);
-            }
-        }
-
-        & .product__content {
-            width: 50%;
-        }
-
-        & .product__summary {
-            justify-content: space-between;
-            align-items: center;
-            width: 100%;
-
-
-            &-price {
-                margin: 0;
-            }
-        }
-    }
-
-    &__pagination {
-        display: flex;
-        justify-content: center;
-        margin-top: 2em;
-        padding: 2em 0;
-        width: 100%;
-        font-size: 1.2rem;
-    }
-}
-
-.pagination-controls::ng-deep .ngx-pagination {
-    padding-left: 0;
-}
-
-.pagination-controls::ng-deep .ngx-pagination .current {
-    background: rgb(223, 197, 201);
-}
-
-@media (min-width: 476px) {
-    .product {
-        width: 45%;
-
-        &.list {
-            width: 100%;
-
-            & .product__img {
-                width: 40%;
-                height: 130px;
-            }
-        }
-    }
-}
-
-@media (min-width: 589px) {
-    .product {
-        width: 45%;
-
-        &.list {
-            width: 100%;
-        }
-    }
-}
-
-@media (min-width: 976px) {
-    .product {
-        padding: 2em;
-        width: 30%;
-
-        &__pagination {
-            margin-top: 3em;
-        }
-
-        &.list {
-            width: 100%;
-
-            & .product__img {
-                padding: 0 0 1em 1em;
-                width: 40%;
-            }
-        }
-
-    }
-}
-
-@media (min-width: 1200px) {
-    .product {
-        width: 23%;
-
-        &.list {
-            width: 80%;
-            max-width: 850px;
-
-            & .product__img {
-                width: 30%;
-                padding: 0 0 1em 1em;
-            }
-        }
-    }
-}
